pyproject.toml
[tool.poetry]
name = "auto-scheduler"
version = "0.1.0"
description = "AI-powered content scheduling and management platform"
authors = ["Your Name <your.email@example.com>"]
readme = "README.md"
packages = [{include = "app"}]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.115.12"
uvicorn = {extras = ["standard"], version = "^0.27.1"}
sqlalchemy = "^2.0.27"
psycopg2-binary = "^2.9.9"
python-jose = {extras = ["cryptography"], version = "^3.4.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
python-multipart = "^0.0.18"
redis = "^5.0.1"
python-dotenv = "^1.0.1"
pydantic = "^2.5.2"
pydantic-settings = "^2.1.0"
httpx = "^0.26.0"
alembic = "^1.13.1"
email-validator = "^2.1.0.post1"
python-magic = "^0.4.27"
celery = "^5.3.6"
prometheus-client = "^0.19.0"
pillow = "^10.3.0"
cryptography = "^44.0.1"
boto3 = "^1.34.0"
openai = "^1.0.0"
sentry-sdk = "^1.40.0"
pytest = "^8.0.0"
pytest-cov = "^4.1.0"
black = "^24.3.0"
flake8 = "^7.0.0"
mypy = "^1.8.0"
isort = "^5.13.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.0.0"
pytest-cov = "^4.1.0"
black = "^24.3.0"
flake8 = "^7.0.0"
mypy = "^1.8.0"
isort = "^5.13.0"

[build-system]
requires = ["setuptools>=42.0.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.black]
line-length = 88
target-version = ['py38']
include = '\.pyi?$'

[tool.isort]
profile = "black"
multi_line_output = 3

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
check_untyped_defs = true

[[tool.mypy.overrides]]
module = ["sqlalchemy.*", "dotenv.*"]
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
addopts = "-v --cov=app --cov-report=term-missing" 